requirements.txt
glfw==2.4.0
igl==2.2.1
numpy==1.21.4
opencv_python==4.5.5.64
PyOpenGL==3.1.0
pytorch3d==0.6.0
PyYAML==5.3.1
scikit_image==0.17.2
scipy==1.8.1
torch==1.8.0
torchvision==0.9.0
trimesh==3.9.15
